# Optimize guid creation under net6. Yes, again.