--- a/Vostok.Commons.Threading/GuidGenerator.cs
+++ b/Vostok.Commons.Threading/GuidGenerator.cs
@@ -1,4 +1,8 @@
 ﻿using System;
+#if NET6_0_OR_GREATER
+using System.Runtime.CompilerServices;
+#else
+#endif
 using JetBrains.Annotations;
 
 namespace Vostok.Commons.Threading
@@ -6,6 +10,10 @@ namespace Vostok.Commons.Threading
     [PublicAPI]
     internal static class GuidGenerator
     {
+#if NET6_0_OR_GREATER
+        [SkipLocalsInit]
+#else
+#endif
         public static unsafe Guid GenerateNotCryptoQualityGuid()
         {
             var bytes = stackalloc byte[16];
@@ -13,11 +21,15 @@ namespace Vostok.Commons.Threading
 
             var random = ThreadSafeRandom.ObtainThreadStaticRandom();
 
+#if NET6_0_OR_GREATER
+            random.NextBytes(new Span<byte>(bytes, 16));
+#else
             for (var i = 0; i < 4; i++)
             {
                 *(int*)dst = random.Next();
                 dst += 4;
             }
+#endif
 
             return *(Guid*)bytes;
         }