--- a/Vostok.Commons.Threading/GuidGenerator.cs
+++ b/Vostok.Commons.Threading/GuidGenerator.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 #if NET6_0_OR_GREATER
 using System.Runtime.CompilerServices;
-#else
 #endif
 using JetBrains.Annotations;
 
@@ -12,26 +11,30 @@ namespace Vostok.Commons.Threading
     {
 #if NET6_0_OR_GREATER
         [SkipLocalsInit]
+        public static unsafe Guid GenerateNotCryptoQualityGuid()
+        {
+            var bytes = stackalloc byte[16];
+            
+            Random.Shared.NextBytes(new Span<byte>(bytes, 16));
+
+            return *(Guid*)bytes;
+        }
 #else
-#endif
         public static unsafe Guid GenerateNotCryptoQualityGuid()
         {
             var bytes = stackalloc byte[16];
             var dst = bytes;
 
             var random = ThreadSafeRandom.ObtainThreadStaticRandom();
-
-#if NET6_0_OR_GREATER
-            random.NextBytes(new Span<byte>(bytes, 16));
-#else
+            
             for (var i = 0; i < 4; i++)
             {
                 *(int*)dst = random.Next();
                 dst += 4;
             }
-#endif
 
             return *(Guid*)bytes;
         }
+#endif
     }
 }