This is the benchmark result:
```
// * Summary *

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1526 (21H2)
Intel Core i7-4771 CPU 3.50GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.103
  [Host]   : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  .Net 4.8 : .NET Framework 4.8 (4.8.4470.0), X64 RyuJIT
  .Net 5.0 : .NET 5.0.15 (5.0.1522.11506), X64 RyuJIT
  .Net 6.0 : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT


|                      Method |            Runtime |     Mean |
|---------------------------- |------------------- |---------:|
|                 GuidNewGuid | .NET Framework 4.8 | 81.36 ns |
|               GenerateInt32 | .NET Framework 4.8 | 56.41 ns |
|               GenerateInt64 | .NET Framework 4.8 | 55.54 ns |
|               GenerateBytes | .NET Framework 4.8 | 56.11 ns |
| GenerateBytesSkipLocalsInit | .NET Framework 4.8 | 55.60 ns |
|---------------------------- |------------------- |---------:|
|                 GuidNewGuid |           .NET 5.0 | 80.87 ns |
|               GenerateInt32 |           .NET 5.0 | 45.13 ns |
|               GenerateInt64 |           .NET 5.0 | 44.93 ns |
|               GenerateBytes |           .NET 5.0 | 45.01 ns |
| GenerateBytesSkipLocalsInit |           .NET 5.0 | 45.42 ns |
|---------------------------- |------------------- |---------:|
|                 GuidNewGuid |           .NET 6.0 | 80.38 ns |
|               GenerateInt32 |           .NET 6.0 | 46.70 ns |
|               GenerateInt64 |           .NET 6.0 | 23.35 ns |
|               GenerateBytes |           .NET 6.0 | 16.76 ns |
| GenerateBytesSkipLocalsInit |           .NET 6.0 | 15.40 ns |
```

GenerateBytesSkipLocalsInit - this is the new implementation.
GenerateInt32 - the previous one
GuidNewGuid - the pre-previous one

Another methods here are just for fun.